test(mitm-addon): split response hook tests
